node2-processor/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
numpy==1.24.3
librosa==0.10.0
torch==2.0.0
psycopg2-binary==2.9.9
twilio==8.10.0
python-dotenv==1.0.0
pydantic==2.5.0